# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/calendars/events": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all events for user's calendars within a specified time range (max 3 months)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calendar"
                ],
                "summary": "Get Calendar Events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start timestamp in Unix format",
                        "name": "start_timestamp",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End timestamp in Unix format",
                        "name": "end_timestamp",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Force sync from Google API regardless of cache",
                        "name": "force_sync",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Events retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/model.CalendarEventsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid query parameters or time range",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Authentication required",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/calendars/google": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all Google calendars for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calendar"
                ],
                "summary": "Get User Calendars",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Force sync from Google API regardless of cache",
                        "name": "force_sync",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calendars retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/model.CalendarListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Authentication required",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - Google token not found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Imports a specific Google calendar to the user's database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calendar"
                ],
                "summary": "Import Google Calendar",
                "parameters": [
                    {
                        "description": "Import calendar request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/calendar.ImportCalendarRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Calendar imported successfully",
                        "schema": {
                            "$ref": "#/definitions/calendar.ImportCalendarResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Authentication required",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - Google token not found or calendar not found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict - Calendar already imported",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "calendar.ImportCalendarRequest": {
            "type": "object",
            "required": [
                "calendar_id"
            ],
            "properties": {
                "calendar_id": {
                    "type": "string"
                }
            }
        },
        "calendar.ImportCalendarResponse": {
            "type": "object",
            "properties": {
                "calendar": {
                    "$ref": "#/definitions/model.Calendar"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "model.Calendar": {
            "description": "Calendar",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "event_color": {
                    "type": "string"
                },
                "event_nickname": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "source": {
                    "$ref": "#/definitions/model.CalendarSource"
                },
                "source_id": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "synced_at": {
                    "type": "string"
                },
                "time_zone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string",
                    "example": "0"
                },
                "visibility": {
                    "$ref": "#/definitions/model.CalendarVisibility"
                }
            }
        },
        "model.CalendarEvent": {
            "description": "Calendar event",
            "type": "object",
            "properties": {
                "allDay": {
                    "description": "True if it's an all-day event",
                    "type": "boolean"
                },
                "calendar_id": {
                    "description": "calendar ID",
                    "type": "string",
                    "example": "0"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "Optional description",
                    "type": "string"
                },
                "end": {
                    "description": "ISO8601 datetime or date",
                    "type": "string"
                },
                "event_color": {
                    "description": "Optional display color",
                    "type": "string"
                },
                "id": {
                    "description": "Google Event ID",
                    "type": "string",
                    "example": "0"
                },
                "location": {
                    "description": "Optional event location",
                    "type": "string"
                },
                "source_id": {
                    "description": "Source calendar ID",
                    "type": "string"
                },
                "start": {
                    "description": "ISO8601 datetime or date (for all-day)",
                    "type": "string"
                },
                "title": {
                    "description": "Event title (summary)",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "visibility": {
                    "description": "public / private / default",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.CalendarEventVisibility"
                        }
                    ]
                }
            }
        },
        "model.CalendarEventVisibility": {
            "type": "string",
            "enum": [
                "public",
                "private",
                "inherited"
            ],
            "x-enum-varnames": [
                "CalendarEventVisibilityPublic",
                "CalendarEventVisibilityPrivate",
                "CalendarEventVisibilityInherited"
            ]
        },
        "model.CalendarEventsResponse": {
            "description": "Calendar events response",
            "type": "object",
            "properties": {
                "calendars": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CalendarWithEvents"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Calendar events retrieved successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "model.CalendarListResponse": {
            "description": "Calendar list response",
            "type": "object",
            "properties": {
                "calendars": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GoogleCalendar"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Calendars retrieved successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "model.CalendarSource": {
            "type": "string",
            "enum": [
                "google",
                "isc"
            ],
            "x-enum-varnames": [
                "SourceGoogle",
                "SourceISC"
            ]
        },
        "model.CalendarVisibility": {
            "type": "string",
            "enum": [
                "public",
                "private"
            ],
            "x-enum-varnames": [
                "CalendarVisibilityPublic",
                "CalendarVisibilityPrivate"
            ]
        },
        "model.CalendarWithEvents": {
            "description": "Calendar with events",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "event_color": {
                    "type": "string"
                },
                "event_nickname": {
                    "type": "string"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CalendarEvent"
                    }
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "source": {
                    "$ref": "#/definitions/model.CalendarSource"
                },
                "source_id": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "synced_at": {
                    "type": "string"
                },
                "time_zone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string",
                    "example": "0"
                },
                "visibility": {
                    "$ref": "#/definitions/model.CalendarVisibility"
                }
            }
        },
        "model.ErrorResponse": {
            "description": "Error response structure",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Detailed error information",
                    "type": "string",
                    "example": "invalid email or password"
                },
                "message": {
                    "description": "Error message",
                    "type": "string",
                    "example": "Authentication failed"
                },
                "success": {
                    "description": "Always false for error responses",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "model.GoogleCalendar": {
            "description": "Google Calendar information",
            "type": "object",
            "properties": {
                "accessRole": {
                    "type": "string",
                    "example": "owner"
                },
                "backgroundColor": {
                    "type": "string",
                    "example": "#9c27b0"
                },
                "colorId": {
                    "type": "string",
                    "example": "1"
                },
                "conferenceProperties": {
                    "$ref": "#/definitions/model.GoogleCalendarConferenceProperties"
                },
                "deleted": {
                    "type": "boolean",
                    "example": false
                },
                "description": {
                    "type": "string",
                    "example": "Personal calendar"
                },
                "etag": {
                    "type": "string",
                    "example": "\"00000000000000000000\""
                },
                "foregroundColor": {
                    "type": "string",
                    "example": "#ffffff"
                },
                "hidden": {
                    "type": "boolean",
                    "example": false
                },
                "id": {
                    "type": "string",
                    "example": "primary"
                },
                "kind": {
                    "type": "string",
                    "example": "calendar#calendarListEntry"
                },
                "location": {
                    "type": "string",
                    "example": "Mountain View, CA"
                },
                "primary": {
                    "type": "boolean",
                    "example": true
                },
                "selected": {
                    "type": "boolean",
                    "example": true
                },
                "summary": {
                    "type": "string",
                    "example": "My Calendar"
                },
                "summaryOverride": {
                    "type": "string",
                    "example": "Custom Summary"
                },
                "timeZone": {
                    "type": "string",
                    "example": "America/Los_Angeles"
                }
            }
        },
        "model.GoogleCalendarConferenceProperties": {
            "description": "Google Calendar conference properties",
            "type": "object",
            "properties": {
                "allowedConferenceSolutionTypes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"hangoutsMeet\"]"
                    ]
                }
            }
        },
...
    }
}